Split pr_inversion.py into an inversion/ subpackage
The 1366-line module had four separable layers, now four modules:
basis.py (sine-basis functions and transforms), solver.py (data
preparation, matrix assembly, SVD solve, invert()), estimate.py
(alpha/n_terms heuristics, auto_invert, D_max scans) and result.py
(PrResult and the error types).

sans_fitter.pr_inversion remains importable as an alias of
sans_fitter.inversion, so existing 'from sans_fitter import
pr_inversion' usage keeps working. Tests that reached module-private
helpers now import the specific submodule.
