+++ src/sans_fitter/__init__.py
@@ -2,12 +2,13 @@
 SANS Model Fitter - A flexible template for fitting SANS data with SasModels
 """
 
-from . import examples, pr_inversion
+from . import examples
+from . import inversion as pr_inversion
 from .data import ops as data_ops
 from .fitter import SANSFitter, get_all_models
+from .inversion import InsufficientDataError, PrEstimationError, PrResult
 from .modeling.parameters import ParameterManager
 from .modeling.polydispersity import PD_DEFAULTS, PD_DISTRIBUTION_TYPES
-from .pr_inversion import InsufficientDataError, PrEstimationError, PrResult
 from .results import FitResultContract, PosteriorSummary
 
 __version__ = '0.3.0'
